% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggtrout_cb}
\alias{ggtrout_cb}
\title{Check whether ggtrout palettes are colorblind friendly
Uses the \link[colorblindcheck]{palette_check} function}
\usage{
ggtrout_cb(name)
}
\arguments{
\item{name}{name of ggtrout palette (e.g., \code{"brook1"})}
}
\value{
values for color vision deficiencies
}
\description{
Check whether ggtrout palettes are colorblind friendly
Uses the \link[colorblindcheck]{palette_check} function
}
