% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggtrout_palette}
\alias{ggtrout_palette}
\title{Generates ggtrout palette}
\usage{
ggtrout_palette(name, n, type = c("discrete", "continuous"), quiet = FALSE)
}
\arguments{
\item{name}{name of palette. Options are: \code{"brook1"}, \code{"brook2"}, \code{"brook3"}, \code{"cutthroat1"}, \code{"cutthroat2"}, \code{"cutthroat3"}, \code{"rainbow1"}, \code{"rainbow2"}, \code{"rainbow3"}, \code{"greenback"}}

\item{n}{number of desired colors in palette; defaults to all colors in given palette.}

\item{type}{either \code{"continuous"} or \code{"discrete"} (default). Using \code{"continuous"} will automatically interpolate between colors.}

\item{quiet}{whether to print palette to screen (defaults to FALSE)}
}
\value{
a vector of colors
}
\description{
This function is largely adapted from the \code{wes_palette()} function in the \href{https://github.com/karthik/wesanderson/tree/master}{wesanderson} package.
}
\examples{
ggtrout_palette(name = "brook1", type = "continuous")
ggtrout_palette(name = "rainbow3", n = 3)

}
