% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ggtrout_palettes}
\alias{ggtrout_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 10.
}
\usage{
ggtrout_palettes
}
\description{
Use \code{\link{ggtrout_palette}} to construct palettes of desired length.
}
\keyword{datasets}
