% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggtrout_type}
\alias{ggtrout_type}
\title{List whether palette is qualitative, diverging, or monochromatic}
\usage{
ggtrout_type(name = "all")
}
\arguments{
\item{name}{list type of given palette (e.g., \code{"brook_div"}), list all palettes of desired type (e.g., \code{"monochromatic"}), or list types of all (\code{"all"}; default)}
}
\value{
prints types of palettes
}
\description{
List whether palette is qualitative, diverging, or monochromatic
}
