% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{grid_palettes}
\alias{grid_palettes}
\title{Grid several palettes for README}
\usage{
grid_palettes(family)
}
\arguments{
\item{family}{trout (sub)species name (options are "brook", "cutthroat" (includes greenback), or "rainbow"; can also do "all" to print all palettes)}
}
\value{
prints ggplots for all palettes within a given family
}
\description{
Grid several palettes for README
}
