% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_trout_example}
\alias{load_trout_example}
\title{Load example datasets}
\usage{
load_trout_example(quiet = FALSE)
}
\arguments{
\item{quiet}{whether to load without messages}
}
\value{
loads chibasin, bluelakes, and animas objects
}
\description{
Load example datasets
}
