% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{plot_helper}
\alias{plot_helper}
\title{Helper function for building palette plots}
\usage{
plot_helper(pal)
}
\arguments{
\item{pal}{palette (output from \code{ggtrout_palette()} function)}
}
\value{
ggplot-type object with palette
}
\description{
Helper function for building palette plots
}
