% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{print_palette}
\alias{print_palette}
\title{Helper function to print palettes}
\usage{
print_palette(x, name)
}
\arguments{
\item{x}{palette}

\item{name}{palette name}
}
\value{
prints palette
}
\description{
Helper function to print palettes
}
